Move the code relevant to Pager's state minimiser into its own module.
This isn't totally clean, but it's reasonably decent I think.
